requirements.txt
httpx>=0.28.0
pyjwt>=2.12.0
python-dotenv>=1.0.0
streamlit>=1.40.0
sentence-transformers>=2.7.0
numpy>=1.26
